Improve control over redirects during auth code flows (#248)
When sheparding a user through an interactive authorisation code flow to
access APIs protected by OAuth2, the oauth_flow_auth_code() function
runs a temporary webserver on localhost to capture the code. This is a
very common technique for OAuth client libraries.

Unfortunately, this doesn't work in environments where localhost is
probably inaccessible to users -- which is normally the case on RStudio
Server and Posit Workbench. Previously the oauth_flow_auth_code()
function would hang indefinitely in these environments.

This commit fixes this failure mode by attempting to detect when this
temporary localhost server will actually work and only runs one if it
will.

Of course, this doesn't resolve the issue for users that can't run a
localhost server -- so in this situation, we now allow users to manually
paste the authorisation code into the console from... somewhere. In the
likely event that they don't want to be redirected to localhost for this
step, this commit also fulfills the longstanding request for control
over the redirect URL.

The combination of a manual copy & paste step with a custom redirect URL
is what the gargle package terms the "pseudo out-of-band" flow, and the
design used here is largely based on that package.

Because of this compatibility, users could register the existing public
page at https://www.tidyverse.org/google-callback/ with their OAuth
provider, host their own clone, or create a variant.

Control over the redirect URL is backwards-compatible but does require
making the 'host_name' parameter deprecated, unfortunately.

Unit tests for new functions are included, as are various updates to the
documentation.

Fixes #165