Refine `resp_stream_sse()` (#652)
It's a bit difficult to translate the spec to the behaviour that we should implement, but I think this bullet:

> If the data buffer is an empty string, set the data buffer and the event type buffer to the empty string and return.

Implies that if the event has no data, we should automatically wait for an event with data. This allows APIs to stream SSEs containing only comments to keep the stream alive. I think it makes sense to handle this hear, rather than  requiring the end user to deal with it.

`event$data` is also now guaranteed to be a single string, which I think is easier to deal with.

Fixes #650