Improved parallel runner (#681)
The queue is now managed in R rather than by curl. This means that we can refresh OAuth tokens, use the cache, respect throttling, and do retries. For simplicity of implementation, throttling and rate-limiting are applied to requests as a whole, making this most suitable for repeated requests to the same API/website.