Avoid partial matching of "refresh_token"
A global search for `\w+\$\w+\s*<-` finds a ton of results. Is there a way to make this safer globally? See #696 for the full description of why this is an issue here, but the same "editing a list that we don't fully own" issue could be important in a lot of other places in this package.

Closes #696.